pyproject.toml
[tool.poetry]
name = "muzlock"
version = "0.1.0"
description = ""
package-mode = false
authors = ["Aleksandr <aleksandr.voitushenko@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.10.0"
python-dotenv = "^1.0.1"
pymongo = "^4.8.0"
mongoengine = "^0.28.2"
certifi = "^2024.7.4"
beautifulsoup4 = "^4.12.3"
requests = "^2.32.3"
lxml = "^5.3.0"
fastapi = "^0.114.2"
uvicorn = "^0.30.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
